--- lib/agent/base.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module Agent
-  class Base
-  end
-end
