+++ lib/langchain/agent/base.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Langchain
+  module Agent
+    class Base
+    end
+  end
+end
