+++ lib/langchain/agent/chain_of_thought_agent/chain_of_thought_agent_prompt.json
