--- spec/agent/base_spec.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-RSpec.describe Agent::Base do
-end
