+++ spec/langchain/agent/base_spec.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+RSpec.describe Langchain::Agent::Base do
+end
